+++ include/Enemy/Enemies/Troll.h
@@ -6,7 +6,7 @@
 class Troll : public Enemy
 {
 public:
-  Troll(short unsigned int playerLevel);
+  Troll(int playerLevel);
 };
 
 #endif // DASTARDLY_DUNGEON_DWELLERS_TROLL_H
