+++ include/Player/Inventory.h
@@ -5,7 +5,7 @@ class Inventory
 {
 private:
   // Inventory Variables
-  short unsigned int gold = 0, potions = 3, armor = 0, keys = 0;
+  int gold = 0, potions = 3, armor = 0, keys = 0;
 
 public:
   Inventory();
@@ -16,21 +16,21 @@ public:
 
 public:
   // Inventory Accessors
-  short unsigned int getGold() { return gold; }
-  short unsigned int getPotions() { return potions; }
-  short unsigned int getArmor() { return armor; }
-  short unsigned int getKeys() { return keys; }
+  int getGold() { return gold; }
+  int getPotions() { return potions; }
+  int getArmor() { return armor; }
+  int getKeys() { return keys; }
 
   // Inventory Mutators
-  void addGold(short unsigned int goldAdjust);
-  void addPotions(short unsigned int potionAdjust);
-  void addArmor(short unsigned int armorAdjust);
-  void addKeys(short unsigned int keyAdjust);
+  void addGold(int goldAdjust);
+  void addPotions(int potionAdjust);
+  void addArmor(int armorAdjust);
+  void addKeys(int keyAdjust);
 
-  void removeGold(short unsigned int goldAdjust) { gold -= goldAdjust; }
-  void removePotions(short unsigned int potionAdjust) { potions -= potionAdjust; }
-  void removeArmor(short unsigned int armorAdjust) { armor -= armorAdjust; }
-  void removeKeys(short unsigned int keysAdjust) { keys -= keysAdjust; }
+  void removeGold(int goldAdjust) { gold -= goldAdjust; }
+  void removePotions(int potionAdjust) { potions -= potionAdjust; }
+  void removeArmor(int armorAdjust) { armor -= armorAdjust; }
+  void removeKeys(int keysAdjust) { keys -= keysAdjust; }
 };
 
 #endif // DASTARDLY_DUNGEON_DWELLERS_INVENTORY_H
